added theme for typography,  singleplayer page layout
