+++ src/CustomThemeProvider.tsx
@@ -8,6 +8,13 @@ function CustomThemeProvider({ children }: CustomThemeProviderProps) {
       primary: { main: grey[800], dark: grey[900], light: grey[700] }, // dark mode
       secondary: { main: grey[300], dark: grey[400], light: grey[200] }, // light mode
     },
+    typography: {
+      fontFamily: "Bungee, sans-serif",
+      subtitle1: {
+        fontFamily: "monospace",
+        fontSize: 18,
+      },
+    },
   });
 
   return <ThemeProvider theme={theme}>{children}</ThemeProvider>;
