--- src/pages/Singleplayer.tsx
@@ -1,9 +1,0 @@
-import Typography from '@mui/material/Typography';
-
-function Singleplayer() {
-    return (
-        <Typography color='primary.main'>SINGLE PLAYER</Typography>
-    )
-}
-
-export default Singleplayer
